refactor: reorganize components into directories
