+++ src/containers/SpotSet.jsx
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 import * as Constants from '../constants/ApiConstants';
-import Error from './Error';
+import Error from './Banners/Error';
 import HttpClient from '../utilities/HttpClient';
-import Login from './Login';
-import Logout from './Logout';
-import Setlist from './Setlist';
-import Setlists from './Setlists';
+import Login from './Sidecard/Login';
+import Logout from './Sidecard/Logout';
+import Setlist from './SearchResults/Setlist';
+import Setlists from './SearchForm/Setlists';
 import SpotifyFunctions from '../helpers/SpotifyFunctions';
 
 export default class SpotSet extends Component {
@@ -23,7 +23,7 @@ export default class SpotSet extends Component {
       error: null,
     }
   }
-  
+
   componentDidMount() {
     this.checkForSetlist();
     this.checkForAccessToken();
@@ -52,13 +52,13 @@ export default class SpotSet extends Component {
 
   clearSetlist = () => {
     localStorage.removeItem('setlist_id');
-    this.setState({ 
+    this.setState({
       setlistId: null,
       playlistUrl: null,
       error: null
     });
   }
-  
+
   isValid = () => {
     let expiresAt = JSON.parse(localStorage.getItem('expires_at'));
     return new Date().getTime() < expiresAt;
@@ -69,8 +69,8 @@ export default class SpotSet extends Component {
     localStorage.removeItem('expires_at');
     localStorage.removeItem('setlist_id');
 
-    this.setState({ 
-      isAuthenticated: false, 
+    this.setState({
+      isAuthenticated: false,
       accessToken: null
     })
   };
@@ -100,7 +100,7 @@ export default class SpotSet extends Component {
         }
 
         { setlistId &&
-          <Setlist 
+          <Setlist
             httpClient={this.httpClient}
             setlistId={setlistId}
             clearSetlist={this.clearSetlist}
@@ -111,7 +111,7 @@ export default class SpotSet extends Component {
         }
 
         <div id="Spotify">
-          {!isAuthenticated ? <Login spotifyFunctions={this.spotifyFunctions} /> : <Logout logOutHandler={this.logout} /> }
+          {!isAuthenticated ? <Login spotifyFunctions={this.spotifyFunctions} /> : <Logout logOutHandler={this.logout} />}
         </div>
 
         { error &&
