+++ src/tests/containers/SearchResults/Track.test.js
@@ -1,11 +1,11 @@
 /* eslint-disable no-undef */
 import { shallow } from 'enzyme';
 import React from 'react';
-import Track from '../../containers/Track';
+import Track from '../../../containers/SearchResults/Track';
 
 describe('Track Component', () => {
   it('sets addedToList to true when addTrack is called', (done) => {
-    const addTrack = Object.assign(jest.fn(), { addTrack: () => {} });
+    const addTrack = Object.assign(jest.fn(), { addTrack: () => { } });
     const sampleUri = 'spotify:track:sampleUri1';
     const wrapper = shallow(<Track uri={'sampleUri'} addTrack={addTrack} />);
 
@@ -19,7 +19,7 @@ describe('Track Component', () => {
   });
 
   it('sets addedToList to false when removeTrack is called', (done) => {
-    const removeTrack = Object.assign(jest.fn(), { removeTrack: () => {} });
+    const removeTrack = Object.assign(jest.fn(), { removeTrack: () => { } });
     const sampleUri = 'spotify:track:sampleUri1';
     const wrapper = shallow(<Track uri={'sampleUri'} removeTrack={removeTrack} />);
