+++ src/tests/containers/Sidecard/Logout.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-undef */
 import { shallow } from 'enzyme';
 import React from 'react';
-import Logout from '../../containers/Logout';
+import Logout from '../../../containers/Sidecard/Logout';
 
 describe('Logout Component', () => {
   it('Log out button calls the logout handler when clicked', () => {
